// src/app/api/models/company-data-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CompanyDataResponse {
  address: string | null;
  address2: string | null;
  companyName: string;
  createdAt: string;
  deletedAt: string | null;
  email: string | null;
  facebookLink: string | null;
  id: string;
  instagramLink: string | null;
  latitude: number | null;
  linkedinLink: string | null;
  longitude: number | null;
  name: string | null;
  phone: string | null;
  signaturePath: string | null;
  taxId: string | null;
  twitterLink: string | null;
  updatedAt: string;
  userId: string;
  website: string | null;
}